+++ crates/usvg/src/parser/mod.rs
@@ -20,9 +20,6 @@ mod use_node;
 #[cfg(feature = "text")]
 mod text;
 
-#[cfg(feature = "text")]
-pub(crate) use converter::Cache;
-
 pub use image::{ImageHrefDataResolverFn, ImageHrefResolver, ImageHrefStringResolverFn};
 pub use options::Options;
 pub(crate) use svgtree::{AId, EId};